S3 Methods for Extracting Factor Scores (Person Classifications)

This S3 method extracts factor scores or skill classifications.

IRT.factor.scores(object, ...)

# S3 method for din
IRT.factor.scores(object, type="MLE", ...)

# S3 method for gdina
IRT.factor.scores(object, type="MLE", ...)

# S3 method for mcdina
IRT.factor.scores(object, type="MLE", ...)

# S3 method for gdm
IRT.factor.scores(object, type="EAP", ...)

# S3 method for slca
IRT.factor.scores(object, type="MLE", ...)

Arguments

object

Object of classes din, gdina, mcdina, gdm or slca.

type

Type of estimated factor score. This can be "MLE", "MAP" or "EAP". The type EAP cannot be used for objects of class slca.

...

More arguments to be passed.

Value

A matrix or a vector with classified scores.

See also

Examples

#############################################################################
# EXAMPLE 1: Extracting factor scores in the DINA model
#############################################################################

data(sim.dina, package="CDM")
data(sim.qmatrix, package="CDM")

# estimate DINA model
mod1 <- CDM::din( sim.dina, q.matrix=sim.qmatrix)
summary(mod1)
# MLE
fsc1a <- CDM::IRT.factor.scores(mod1)
# MAP
fsc1b <- CDM::IRT.factor.scores(mod1, type="MAP")
# EAP
fsc1c <- CDM::IRT.factor.scores(mod1, type="EAP")
# compare classification for skill 1
stats::xtabs( ~ fsc1a[,1] + fsc1b[,1] )
graphics::boxplot( fsc1c[,1] ~ fsc1a[,1] )