S3 Methods for Extracting Item Response Functions

This S3 method extracts item response functions evaluated at a grid of abilities (skills). Item response functions can be plotted using the IRT.irfprobPlot function.

IRT.irfprob(object, ...)

# S3 method for din
IRT.irfprob(object, ...)

# S3 method for gdina
IRT.irfprob(object, ...)

# S3 method for gdm
IRT.irfprob(object, ...)

# S3 method for mcdina
IRT.irfprob(object, ...)

# S3 method for reglca
IRT.irfprob(object, ...)

# S3 method for slca
IRT.irfprob(object, ...)

Arguments

object

Object of classes din, gdina, mcdina, gdm, slca, reglca.

...

More arguments to be passed.

Value

An array with item response probabilities (items \(\times\)

categories \(\times\) skill classes [\(\times\) group]) and attributes

theta

Uni- or multidimensional skill space (theta grid in item response models).

prob.theta

Probability distribution of theta

skillspace

Design matrix and estimated parameters for skill space distribution (only for IRT.posterior.slca)

G

Number of groups

See also

Examples

if (FALSE) {
#############################################################################
# EXAMPLE 1: Extracting item response functions mcdina model
#############################################################################

data(data.cdm02, package="CDM")

dat <- data.cdm02$data
q.matrix <- data.cdm02$q.matrix

#-- estimate model
mod1 <- CDM::mcdina( dat, q.matrix=q.matrix)
#-- extract item response functions
prmod1 <- CDM::IRT.irfprob(mod1)
str(prmod1)
}