Extract Estimated Item Parameters and Skill Class Distribution Parameters

Extracts the estimated parameters from either din, gdina, gdina or gdm objects.

# S3 method for din
coef(object, ...)

# S3 method for gdina
coef(object, ...)

# S3 method for mcdina
coef(object, ...)

# S3 method for gdm
coef(object, ...)

# S3 method for slca
coef(object, ...)

Arguments

object
An object inheriting from either class din, class gdina, class mcdina, class slca or class gdm.

...
Additional arguments to be passed.

Value

A vector, a matrix or a data frame of the estimated parameters for the fitted model.

See also

Examples

data(sim.dina, package="CDM")
data(sim.qmatrix, package="CDM")

# DINA model
d1 <- CDM::din( sim.dina, q.matrix=sim.qmatrix)
coef(d1)

if (FALSE) {
# GDINA model
d2 <- CDM::gdina( sim.dina, q.matrix=sim.qmatrix)
coef(d2)

# GDM model
theta.k <- seq(-4,4,len=11)
d3 <- CDM::gdm( sim.dina, irtmodel="2PL", theta.k=theta.k,
            Qmatrix=as.matrix(sim.qmatrix),  centered.latent=TRUE)
coef(d3)
}