PISA 2000 Reading Study (Chen & de la Torre, 2014)

This is a sub-dataset of the PISA 2000 of German students including 26 items of the reading test. The 26 items was analyzed in Chen and de la Torre (2014) and a subset of 20 items was analyzed in Chen and Chen (2016).

data(data.pisa00R.ct)
data(data.pisa00R.cc)

Format

The format of the dataset data.pisa00R.ct (Chen & de la Torre, 2014) is:

List of 3
$ data :'data.frame': 1095 obs. of 111 variables:
.. [list output truncated]
$ q.matrix: num [1:26, 1:8] 0 1 0 0 0 1 0 0 0 1 ...
..- attr(*, "dimnames")=List of 2
$ skills : chr [1:8] "Locating information" ...

The format of the dataset data.pisa00R.cc (Q-matrix in Chen and Chen, 2016)

List of 2
$ q.matrix:'data.frame': 20 obs. of 5 variables:
..$ A1: num [1:20] 1 1 0 0 1 1 1 0 0 0 ...
..$ A2: num [1:20] 0 0 0 1 0 1 1 1 1 1 ...
..$ A3: num [1:20] 1 1 0 1 1 0 1 0 1 0 ...
..$ A4: num [1:20] 0 1 1 1 0 0 0 0 0 0 ...
..$ A5: num [1:20] 0 0 1 0 0 0 0 1 0 1 ...
$ skills : Named chr [1:5] "Identifying Explicit Information" ...
..- attr(*, "names")=chr [1:5] "A1" "A2" "A3" "A4" ...

References

Chen, H., & Chen, J. (2016). Exploring reading comprehension skill relationships through the G-DINA model. Educational Psychology, 36(6), 1049-1064.

Chen, J., & de la Torre, J. (2014). A procedure for diagnostically modeling extant large-scale assessment data: the case of the programme for international student assessment in reading. Psychology, 5(18), 1967-1978.

Examples

# NOT RUN {
#############################################################################
# EXAMPLE 1: PISA items from Chen and de la Torre (2014)
#            dichotomize item responses
#############################################################################

data(data.pisa00R.ct, package="CDM")
dat <- data.pisa00R.ct$data
Q <- data.pisa00R.ct$q.matrix
resp <- dat[, rownames(Q)]

#** extract item-wise maximum
maxK <- apply( resp, 2, max, na.rm=TRUE )
#** dichotomize response data
resp1 <- resp
for (ii in seq(1,ncol(resp)) ){
    resp1[,ii] <- 1 * ( resp[,ii]==maxK[ii] )
}
# }