Discrimination Indices at Item-Attribute, Item and Test Level

Computes discrimination indices at the probability metric (de la Torre, 2008; Henson, DiBello & Stout, 2018).

discrim.index(object, ...)

# S3 method for din
discrim.index(object, ...)

# S3 method for gdina
discrim.index(object, ...)

# S3 method for mcdina
discrim.index(object, ...)

# S3 method for discrim.index
summary(object, file=NULL, digits=3,  ...)

Arguments

object
Object of class din or gdina.

file
Optional file name for a file in which the summary output should be sunk

digits
Number of digits for rounding

...
Further arguments to be passed

Details

If item \(j\) possesses \(H_j\) categories, the item-attribute specific discrimination for attribute \(k\) according to Henson et al. (2018) is defined as $$ DI_{jk}=\frac{1}{2} \max_{ \boldmath{\alpha} } \left( \sum_{h=1}^{H_j} | P(X_j=h| \boldmath{\alpha} ) - P(X_j=h| \boldmath{\alpha}^{(-k)} ) | \right ) $$ where \(\boldmath{\alpha}^{(-k)}\) and \(\boldmath{\alpha}\) differ only in attribute \(k\). The index \(DI_{jk}\) can be found as the value discrim_item_attribute. The test-level discrimination index is defined as $$\overline{DI}=\frac{1}{J} \sum_{j=1}^J \max_k DI_{jk} $$ and can be found in discrim_test.

According to de la Torre (2008) and de la Torre, Rossi and van der Ark (2018), the item discrimination index (IDI) is defined as $$IDI_j=\max_{ \boldmath{\alpha}_1,\boldmath{\alpha}_2, h} | P(X_j=h| \boldmath{\alpha}_1 ) - P(X_j=h| \boldmath{\alpha}_2 ) |$$ and can be found as idi in the values list.

Value

A list with following entries

discrim_item_attribute

Discrimination indices \(DI_{jk}\) at item level for each attribute

idi

Item discrimination index \(IDI_j\)

discrim_test

Discrimination index at test level

References

de la Torre, J. (2008). An empirically based method of Q-matrix validation for the DINA model: Development and applications. Journal of Educational Measurement, 45, 343-362.
http://dx.doi.org/10.1111/j.1745-3984.2008.00069.x

de la Torre, J., van der Ark, L. A., & Rossi, G. (2018). Analysis of clinical data from a cognitive diagnosis modeling framework. Measurement and Evaluation in Counseling and Development, 51(4), 281-296. https://doi.org/10.1080/07481756.2017.1327286

Henson, R., DiBello, L., & Stout, B. (2018). A generalized approach to defining item discrimination for DCMs. Measurement: Interdisciplinary Research and Perspectives, 16(1), 18-29.
http://dx.doi.org/10.1080/15366367.2018.1436855

See also

Examples

if (FALSE) {
#############################################################################
# EXAMPLE 1: DINA and GDINA model
#############################################################################

data(sim.dina, package="CDM")
data(sim.qmatrix, package="CDM")

#-- fit GDINA and DINA model
mod1 <- CDM::gdina( sim.dina, q.matrix=sim.qmatrix )
mod2 <- CDM::din( sim.dina, q.matrix=sim.qmatrix )

#-- compute discrimination indices
dimod1 <- CDM::discrim.index(mod1)
dimod2 <- CDM::discrim.index(mod2)
summary(dimod1)
summary(dimod2)
}