Fraction Subtraction Q-Matrix

The Q-Matrix corresponding to Tatsuoka (1984) fraction subtraction data set.

data(fraction.subtraction.qmatrix)

Format

The fraction.subtraction.qmatrix data frame consists of \(J=20\) rows and \(K=8\) columns, specifying the attributes that are believed to be involved in solving the items. Each row in the data frame represents an item and the entries in the row indicate whether an attribute is needed to master the item (denoted by a "1") or not (denoted by a "0"). The attributes for the fraction subtraction data set are the following:

alpha1

convert a whole number to a fraction,

alpha2

separate a whole number from a fraction,

alpha3

simplify before subtracting,

alpha4

find a common denominator,

alpha5

borrow from whole number part,

alpha6

column borrow to subtract the second numerator from the first,

alpha7

subtract numerators,

alpha8

reduce answers to simplest form.

Details

This Q-matrix can be found in DeCarlo (2011). It is the same used by de la Torre and Douglas (2004).

Source

DeCarlo, L. T. (2011). On the analysis of fraction subtraction data: The DINA Model, classification, latent class sizes, and the Q-Matrix. Applied Psychological Measurement, 35, 8--26.

References

de la Torre, J. and Douglas, J. (2004). Higher-order latent trait models for cognitive diagnosis. Psychometrika, 69, 333--353.

Tatsuoka, C. (2002). Data analytic methods for latent partially ordered classification models. Journal of the Royal Statistical Society, Series C, Applied Statistics, 51, 337--350.

Tatsuoka, K. (1984) Analysis of errors in fraction addition and subtraction problems. Final Report for NIE-G-81-0002, University of Illinois, Urbana-Champaign.