RMSEA Item Fit

This function estimates a chi squared based measure of item fit in cognitive diagnosis models similar to the RMSEA itemfit implemented in mdltm (von Davier, 2005; cited in Kunina-Habenicht, Rupp & Wilhelm, 2009).

The RMSEA statistic is also called as the RMSD statistic, see IRT.RMSD.

itemfit.rmsea(n.ik, pi.k, probs, itemnames=NULL)

Arguments

n.ik

An array of four dimensions: Classes x items x categories x groups

pi.k

An array of two dimensions: Classes x groups

probs

An array of three dimensions: Classes x items x categories

itemnames

An optional vector of item names. Default is NULL.

Details

For item \(j\), the RMSEA itemfit in this function is calculated as follows: $$ RMSEA_j=\sqrt{ \sum_k \sum_c \pi ( \bold{\theta}_c) \left( P_j ( \bold{\theta}_c ) - \frac{n_{jkc}}{N_{jc}} \right)^2 } $$ where \(c\) denotes the class of the skill vector \(\bold{\theta}\), \(k\) is the item category, \(\pi ( \bold{\theta}_c)\) is the estimated class probability of \(\bold{\theta}_c\), \(P_j\) is the estimated item response function, \(n_{jkc}\) is the expected number of students with skill \(\bold{\theta}_c\) on item \(j\) in category \(k\) and \(N_{jc}\) is the expected number of students with skill \(\bold{\theta}_c\) on item \(j\).

Value

A list with two entries:

rmsea

Vector of RMSEA item statistics

rmsea.groups

Matrix of group-wise RMSEA item statistics

References

Kunina-Habenicht, O., Rupp, A. A., & Wilhelm, O. (2009). A practical illustration of multidimensional diagnostic skills profiling: Comparing results from confirmatory factor analysis and diagnostic classification models. Studies in Educational Evaluation, 35, 64--70.

von Davier, M. (2005). A general diagnostic model applied to language testing data. ETS Research Report RR-05-16. ETS, Princeton, NJ: ETS.

See also