Extract Log-Likelihood

Extracts the log-likelihood from either din, gdina, mcdina, slca or gdm objects.

# S3 method for din
logLik(object, ...)

# S3 method for gdina
logLik(object, ...)

# S3 method for mcdina
logLik(object, ...)

# S3 method for gdm
logLik(object, ...)

# S3 method for slca
logLik(object, ...)

# S3 method for reglca
logLik(object, ...)

Arguments

object

An object inheriting from either class din, gdina, slca, reglca or gdm.

...

Additional arguments

See also

Examples

data(sim.dina, package="CDM")
data(sim.qmatrix, package="CDM")

# logLik method | DINA model
d1 <- CDM::din( sim.dina, q.matrix=sim.qmatrix, rule="DINA")
summary(d1)
lld1 <- logLik(d1)
  ##   > lld1
  ##   'log Lik.' -2042.378 (df=25)
  ##   > attr(lld1,"df")
  ##   [1] 25
  ##   > attr(lld1,"nobs")
  ##   [1] 400
nobs(lld1)

# AIC and BIC
AIC(lld1)
BIC(lld1)