Tetrachoric or Polychoric Correlations between Attributes

This function takes the results of din or gdina and computes tetrachoric or polychoric correlations between attributes (see e.g. Templin & Henson, 2006).

# tetrachoric correlations
skill.cor(object)

# polychoric correlations
skill.polychor(object, colindex=1)

Arguments

object
Object of class din or gdina

colindex
Index which can used for group-wise calculation of polychoric correlations

Value

A list with following entries:

conttable.skills

Bivariate contingency table of all skill pairs

cor.skills

Tetrachoric correlation matrix for skill distribution

References

Templin, J., & Henson, R. (2006). Measurement of psychological disorders using cognitive diagnosis models. Psychological Methods, 11, 287-305.

Examples

data(sim.dino, package="CDM")
data(sim.qmatrix, package="CDM")

# estimate model
d4 <- CDM::din( sim.dino, q.matrix=sim.qmatrix)
# compute tetrachoric correlations
CDM::skill.cor(d4)
  ## estimated tetrachoric correlations
  ##   $cor.skills
  ##             V1        V2        V3
  ##   V1 1.0000000 0.2567718 0.2552958
  ##   V2 0.2567718 1.0000000 0.9842188
  ##   V3 0.2552958 0.9842188 1.0000000