Summary Method for Objects of Class din

S3 method to summarize objects of the class din.

# S3 method for din
summary(object, top.n.skill.classes=6, overwrite=FALSE, ...)

Arguments

object

A required object of class din, obtained from a call to the function din.

top.n.skill.classes

A numeric, specifying the number of skill classes, starting with the most frequent, to be returned. Default value is 6.

overwrite

An optional boolean, specifying wether or not the method is supposed to overwrite an existing log.file. If the log.file exists and overwrite is FALSE, the user is asked to confirm the overwriting.

...

Optional parameters to be passed to or from other methods will be ignored.

Details

The function summary.din returns an object of the class summary.din (see ‘Value’), for which a print method, print.summary.din, is provided. Specific summary information details such as individual item parameters and their discrimination indices can be accessed through assignment (see ‘Examples’).

Value

If the argument object is of required type,

summary.din returns a named list, of the class

summary.din, consisting of the following seven components:

CALL

A character specifying the model rule, the number of items and the number of attributes underlying the items.

IDI

A matrix giving the item discrimination index (IDI; Lee, de la Torre & Park, 2012) for each item \(j\) $$ IDI_j=1 - s_j - g_j, $$ where a high IDI corresponds to favorable test items which have both low guessing and slipping rates.

SKILL.CLASSES

A vector giving the top.n.skill.classes most frequent skill classes and the corresponding class probability.

AIC

A numeric giving the AIC of the specified model object.

BIC

A numeric giving the BIC of the specified model object.

log.file

A character giving the path and file of a specified log file.

din.object

The object of class din for which the summary was requested.

References

Lee, Y.-S., de la Torre, J., & Park, Y. S. (2012). Relationships between cognitive diagnosis, CTT, and IRT indices: An empirical investigation. Asia Pacific Educational Research, 13, 333-345.

Rupp, A. A., Templin, J. L., & Henson, R. A. (2010) Diagnostic Measurement: Theory, Methods, and Applications. New York: The Guilford Press.

See also

Examples

##
## (1) examples based on dataset fractions.subtraction.data
##

## Parameter estimation of DINA model
# rule="DINA" is default
fractions.dina <- CDM::din(data=CDM::fraction.subtraction.data,
           q.matrix=CDM::fraction.subtraction.qmatrix, rule="DINA")

## corresponding summaries, including diagnostic accuracies,
## most frequent skill classes and information
## criteria AIC and BIC
summary(fractions.dina)

## In particular, accessing detailed summary through assignment
detailed.summary.fs <- summary(fractions.dina)
str(detailed.summary.fs)