Artificial Data: DINA and DINO

Artificial data: dichotomously coded fictitious answers of 400 respondents to 9 items assuming 3 underlying attributes.

data(sim.dina)
  data(sim.dino)
  data(sim.qmatrix)

Format

The sim.dina and sim.dino data sets include dichotomous answers of \(N=400\) respondents to \(J=9\) items, thus they are \(400 \times 9\) data matrices. For both data sets \(K=3\) attributes are assumed to underlie the process of responding, stored in sim.qmatrix.

The sim.dina data set is simulated according to the DINA condensation rule, whereas the sim.dino data set is simulated according to the DINO condensation rule. The slipping errors for the items 1 to 9 in both data sets are 0.20, 0.20, 0.20, 0.20, 0.00, 0.50, 0.50, 0.10, 0.03 and the guessing errors are 0.10, 0.125, 0.15, 0.175, 0.2, 0.225, 0.25, 0.275, 0.3. The attributes are assumed to be mastered with expected probabilities of -0.4, 0.2, 0.6, respectively. The correlation of the attributes is 0.3 for attributes 1 and 2, 0.4 for attributes 1 and 3 and 0.1 for attributes 2 and 3.

References

Rupp, A. A., Templin, J. L., & Henson, R. A. (2010) Diagnostic Measurement: Theory, Methods, and Applications. New York: The Guilford Press.

Example Index

Dataset sim.dina

anova (Examples 1, 2), cdi.kli (Example 1), din (Examples 2, 4, 5), gdina (Example 1), itemfit.sx2 (Example 2), modelfit.cor.din (Example 1)

Dataset sim.dino

cdm.est.class.accuracy (Example 1), din (Example 3), gdina (Examples 2, 3, 4),