Root Mean Square Deviation (RMSD) Item Fit Statistic

Computed the item fit statistics root mean square deviation (RMSD), mean absolute deviation (MAD) and mean deviation (MD). See Oliveri and von Davier (2011) for details.

The RMSD statistics was denoted as the RMSEA statistic in older publications, see itemfit.rmsea.

If multiple groups are defined in the model object, a weighted item fit statistic (WRMSD; Yamamoto, Khorramdel, & von Davier, 2013; von Davier, Weeks, Chen, Allen & van der Velden, 2013) is additionally computed.

IRT.RMSD(object)

# S3 method for IRT.RMSD
summary(object, file=NULL, digits=3, ...)

## core computation function
IRT_RMSD_calc_rmsd( n.ik, pi.k, probs, eps=1E-30 )

Arguments

object

Object for which the methods IRT.expectedCounts and IRT.irfprob can be applied.

n.ik

Expected counts

pi.k

Probabilities trait distribution

probs

Item response probabilities

eps

Numerical constant avoiding division by zero

digits

Number of digits used for rounding

file

Optional file name for a file in which summary should be sinked.

...

Optional parameters to be passed.

Details

The RMSD and MD statistics are in operational use in PISA studies since PISA 2015. These fit statistics can also be used for investigating uniform and nonuniform differential item functioning.

Value

List with entries

RMSD

Item-wise and group-wise RMSD statistic

RMSD_bc

Item-wise and group-wise RMSD statistic with analytical bias correction

MAD

Item-wise and group-wise MAD statistic

MD

Item-wise and group-wise MD statistic

chisquare_stat

Item-wise and group-wise \(\chi^2\) statistic

...

Further values

References

Oliveri, M. E., & von Davier, M. (2011). Investigation of model fit and score scale comparability in international assessments. Psychological Test and Assessment Modeling, 53, 315-333.

von Davier, M., Weeks, J., Chen, H., Allen, J., & van der Velden, R. (2013). Creating simple and complex derived variables and validation of background questionnaire data. In OECD (Eds.). Technical Report of the Survey of Adults Skills (PIAAC) (Ch. 20). Paris: OECD.

Yamamoto, K., Khorramdel, L., & von Davier, M. (2013). Scaling PIAAC cognitive data. In OECD (Eds.). Technical Report of the Survey of Adults Skills (PIAAC) (Ch. 17). Paris: OECD.

See also

Examples

if (FALSE) {
#############################################################################
# EXAMPLE 1: data.read | 1PL model in TAM
#############################################################################

data(data.read, package="sirt")
dat <- data.read

#*** Model 1: 1PL model
mod1 <- TAM::tam.mml( resp=dat )
summary(mod1)

# item fit statistics
imod1 <- CDM::IRT.RMSD(mod1)
summary(imod1)

#############################################################################
# EXAMPLE 2: data.math| RMSD and MD statistic for assessing DIF
#############################################################################

data(data.math, package="sirt")
dat <- data.math$data
items <- grep("M[A-Z]", colnames(dat), value=TRUE )

#-- fit multiple group Rasch model
mod <- TAM::tam.mml( dat[,items], group=dat$female )
summary(mod)

#-- fit statistics
rmod <- CDM::IRT.RMSD(mod)
summary(rmod)

#############################################################################
# EXAMPLE 3: RMSD statistic DINA model
#############################################################################

data(sim.dina)
data(sim.qmatrix)
dat <- sim.dina
Q <- sim.qmatrix

#-- fit DINA model
mod1 <- CDM::gdina( dat, q.matrix=Q, rule="DINA" )
summary(mod1)

#-- compute RMSD fit statistic
rmod1 <- CDM::IRT.RMSD(mod1)
summary(rmod1)
}