S3 Method for Extracting Expected Counts

This S3 method extracts expected counts from model output.

IRT.expectedCounts(object, ...)

# S3 method for din
IRT.expectedCounts(object, ...)

# S3 method for gdina
IRT.expectedCounts(object, ...)

# S3 method for gdm
IRT.expectedCounts(object, ...)

# S3 method for mcdina
IRT.expectedCounts(object, ...)

# S3 method for slca
IRT.expectedCounts(object, ...)

# S3 method for reglca
IRT.expectedCounts(object, ...)

Arguments

object
Object of classes din, gdina, mcdina, gdm or slca.

...
More arguments to be passed.

Value

An array with expected counts. The dimensions are items, categories, latent classes and groups.

Examples

if (FALSE) {
#############################################################################
# EXAMPLE 1: Expected counts gdm function
#############################################################################

data(data.fraction1, package="CDM")
dat <- data.fraction1$data
theta.k <- seq( -6, 6, len=11 )   # discretized ability

#--- Model 1: Rasch model
mod1 <- CDM::gdm( dat, irtmodel="1PL", theta.k=theta.k, skillspace="normal",
               centered.latent=TRUE )
emod1 <- CDM::IRT.expectedCounts(mod1)
str(emod1)

#############################################################################
# EXAMPLE 2: Expected counts gdina function
#############################################################################

data(sim.dina, package="CDM")
data(sim.qmatrix, package="CDM")

#--- Model 1: estimation of the GDINA model
mod1 <- CDM::gdina( data=sim.dina, q.matrix=sim.qmatrix)
summary(mod1)
emod1 <- CDM::IRT.expectedCounts(mod1)
str(emod1)

#--- Model 2: GDINA model with two groups
mod2 <- CDM::gdina( data=CDM::sim.dina, q.matrix=CDM::sim.qmatrix,
                   group=rep(1:2, each=200) )
summary(mod2)
emod2 <- CDM::IRT.expectedCounts( mod2 )
str(emod2)
}