Fraction Subtraction Dataset with Different Subsets of Data and Different Q-Matrices

Contains different sub-datasets of the fraction subtraction data of Tatsuoka with different Q-matrix specifications.

data(data.fraction1)
data(data.fraction2)
data(data.fraction3)
data(data.fraction4)
data(data.fraction5)

Format

The dataset data.fraction1 is the fraction subtraction data set with 536 students and 15 items. The Q-matrix was defined in de la Torre (2009). This dataset is a list with the dataset (data) and the Q-matrix as entries.

The format is:

List of 2
$ data :'data.frame':
..$ T01: int [1:536] 0 1 1 1 0 0 0 0 0 0 ...
..$ T02: int [1:536] 1 1 1 1 1 0 0 1 0 0 ...
..$ T03: int [1:536] 0 1 1 1 1 1 0 0 0 0 ...
..$ T04: int [1:536] 1 1 1 0 0 0 0 0 0 0 ...
..$ T05: int [1:536] 0 1 0 0 0 1 1 0 1 1 ...
..$ T06: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
..$ T07: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
..$ T08: int [1:536] 1 1 0 1 1 0 0 0 1 1 ...
..$ T09: int [1:536] 1 1 1 1 0 1 0 0 1 0 ...
..$ T10: int [1:536] 1 1 1 0 0 0 0 0 0 0 ...
..$ T11: int [1:536] 1 1 1 1 0 0 0 0 0 0 ...
..$ T12: int [1:536] 0 1 0 0 0 0 0 0 0 0 ...
..$ T13: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
..$ T14: int [1:536] 1 1 0 0 0 0 0 0 0 0 ...
..$ T15: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
$ q.matrix: int [1:15, 1:5] 1 1 1 1 0 1 1 1 1 1 ...
..- attr(*, "dimnames")=List of 2
.. ..$ : chr [1:15] "T01" "T02" "T03" "T04" ...
.. ..$ : chr [1:5] "QT1" "QT2" "QT3" "QT4" ...

The dataset data.fraction2 is the fraction subtraction data set with 536 students and 11 items. For this data set, several \(Q\) matrices are available. The data is a list. The first entry data contains the data frame. The entry q.matrix1 contains the Q-matrix of Henson, Templin and Willse (2009). The third entry q.matrix2 is an alternative Q-matrix of de la Torre (2009). The fourth entry is a modified Q-matrix of q.matrix1.

The format is:

$ data :'data.frame':
..$ H01: int [1:536] 1 1 1 1 1 0 0 1 0 0 ...
..$ H02: int [1:536] 1 1 1 0 0 0 0 0 0 0 ...
..$ H03: int [1:536] 0 1 0 0 0 1 1 0 1 1 ...
..$ H04: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
..$ H05: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
..$ H06: int [1:536] 1 1 0 1 1 0 0 0 1 1 ...
..$ H08: int [1:536] 1 1 1 0 0 0 0 0 0 0 ...
..$ H09: int [1:536] 1 1 1 1 0 0 0 0 0 0 ...
..$ H10: int [1:536] 0 1 0 0 0 0 0 0 0 0 ...
..$ H11: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
..$ H13: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
$ q.matrix1: int [1:11, 1:3] 1 1 1 1 1 1 1 1 1 1 ...
..- attr(*, "dimnames")=List of 2
.. ..$ : chr [1:11] "H01" "H02" "H03" "H04" ...
.. ..$ : chr [1:3] "QH1" "QH2" "QH3"
$ q.matrix2: int [1:11, 1:5] 1 1 0 1 1 1 1 1 1 1 ...
..- attr(*, "dimnames")=List of 2
.. ..$ : chr [1:11] "H01" "H02" "H03" "H04" ...
.. ..$ : chr [1:5] "QT1" "QT2" "QT3" "QT4" ...
$ q.matrix3: num [1:11, 1:3] 0 0 0 1 0 0 0 0 1 1 ...
..- attr(*, "dimnames")=List of 2
.. ..$ : chr [1:11] "H01" "H02" "H03" "H04" ...
.. ..$ : chr [1:3] "Dim1" "Dim2" "Dim3"

The dataset data.fraction3 contains 12 items and was used in de la Torre (2011).

List of 2
$ data :'data.frame': 536 obs. of 12 variables:
..$ B01: int [1:536] 0 1 1 1 0 0 0 0 0 0 ...
..$ B02: int [1:536] 1 1 1 1 1 0 0 1 0 0 ...
..$ B03: int [1:536] 0 1 1 1 1 1 0 0 0 0 ...
..$ B04: int [1:536] 0 1 0 0 0 1 1 0 1 1 ...
..$ B05: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
..$ B06: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
..$ B07: int [1:536] 1 1 0 1 1 0 0 0 1 1 ...
..$ B08: int [1:536] 1 1 1 1 0 1 0 0 1 0 ...
..$ B09: int [1:536] 1 1 1 1 0 0 0 0 0 0 ...
..$ B10: int [1:536] 0 1 0 0 0 0 0 0 0 0 ...
..$ B11: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
..$ B12: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
$ q.matrix:'data.frame': 12 obs. of 5 variables:
..$ item: Factor w/ 13 levels "","B01","B02",..: 2 3 4 5 6 7 8 9 10 11 ...
..$ QA1 : int [1:12] 1 1 1 1 1 1 1 1 1 1 ...
..$ QA2 : int [1:12] 0 1 0 0 1 1 1 0 0 0 ...
..$ QA3 : int [1:12] 0 1 0 1 1 1 0 1 1 1 ...
..$ QA4 : int [1:12] 0 1 0 0 1 1 0 0 0 1 ...

The dataset data.fraction4 contains 17 items and was used in de la Torre and Douglas (2004) and Chen, Liu, Xu and Ying (2015).

List of 2
$ data :'data.frame': 536 obs. of 17 variables:
..$ A01: int [1:536] 0 0 0 1 0 0 0 0 0 0 ...
..$ A02: int [1:536] 0 1 1 1 0 0 0 0 0 0 ...
..$ A03: int [1:536] 0 1 1 1 0 0 0 0 0 0 ...
..$ A04: int [1:536] 1 1 1 1 1 0 0 1 0 0 ...
..$ A05: int [1:536] 1 1 0 1 1 0 0 0 1 1 ...
..$ A06: int [1:536] 1 1 1 1 0 1 0 0 1 0 ...
..$ A07: int [1:536] 1 1 1 1 0 0 0 0 0 0 ...
..$ A08: int [1:536] 0 0 0 1 0 0 0 0 0 1 ...
..$ A09: int [1:536] 1 1 1 0 0 0 0 0 0 0 ...
..$ A10: int [1:536] 1 1 1 0 0 0 0 0 0 0 ...
..$ A11: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
..$ A12: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
..$ A13: int [1:536] 0 1 0 0 0 0 0 0 0 0 ...
..$ A14: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
..$ A15: int [1:536] 1 1 0 0 0 0 0 0 0 0 ...
..$ A16: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
..$ A17: int [1:536] 0 1 0 0 0 0 0 0 0 0 ...
$ q.matrix:'data.frame': 17 obs. of 9 variables:
..$ item: Factor w/ 18 levels "","A01","A02",..: 2 3 4 5 6 7 8 9 10 11 ...
..$ QA1 : int [1:17] 0 0 0 0 0 0 0 0 1 0 ...
..$ QA2 : int [1:17] 0 0 0 1 0 1 1 1 1 1 ...
..$ QA3 : int [1:17] 0 0 0 1 0 0 0 0 0 0 ...
..$ QA4 : int [1:17] 1 1 1 0 0 0 0 1 0 0 ...
..$ QA5 : int [1:17] 0 0 0 1 0 0 1 0 0 1 ...
..$ QA6 : int [1:17] 1 0 0 0 0 0 1 0 0 0 ...
..$ QA7 : int [1:17] 1 1 1 1 1 1 1 1 1 1 ...
..$ QA8 : int [1:17] 0 0 0 0 1 0 0 1 0 0 ...

The dataset data.fraction5 contains 15 items and was used as an example for the multiple strategy DINA model in de la Torre and Douglas (2008) and Hou and de la Torre (2014). The two Q-matrices for coding the multiple strategies are contained in one matrix q.matrix by joining the columns of both matrices.

List of 2
$ data :'data.frame': 536 obs. of 15 variables:
..$ T01: int [1:536] 0 1 1 1 0 0 0 0 0 0 ...
..$ T02: int [1:536] 1 1 1 1 1 0 0 1 0 0 ...
..$ T03: int [1:536] 0 1 1 1 1 1 0 0 0 0 ...
..$ T04: int [1:536] 1 1 1 0 0 0 0 0 0 0 ...
..$ T05: int [1:536] 0 1 0 0 0 1 1 0 1 1 ...
..$ T06: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
..$ T07: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
..$ T08: int [1:536] 1 1 0 1 1 0 0 0 1 1 ...
..$ T09: int [1:536] 1 1 1 1 0 1 0 0 1 0 ...
..$ T10: int [1:536] 1 1 1 0 0 0 0 0 0 0 ...
..$ T11: int [1:536] 1 1 1 1 0 0 0 0 0 0 ...
..$ T12: int [1:536] 0 1 0 0 0 0 0 0 0 0 ...
..$ T13: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
..$ T14: int [1:536] 1 1 0 0 0 0 0 0 0 0 ...
..$ T15: int [1:536] 1 1 0 1 0 0 0 0 0 0 ...
$ q.matrix:'data.frame': 15 obs. of 15 variables:
..$ item: Factor w/ 16 levels "","T01","T02",..: 2 3 4 5 6 7 8 9 10 11 ...
..$ SA1 : int [1:15] 0 1 1 1 0 1 1 1 1 1 ...
..$ SA2 : int [1:15] 0 1 0 1 0 1 1 1 0 0 ...
..$ SA3 : int [1:15] 0 1 0 1 1 1 1 0 1 1 ...
..$ SA4 : int [1:15] 0 1 0 1 0 1 1 0 0 1 ...
..$ SA5 : int [1:15] 0 0 0 1 0 0 0 0 0 1 ...
..$ SA6 : int [1:15] 0 0 0 0 0 0 0 0 0 0 ...
..$ SA7 : int [1:15] 0 0 0 0 0 0 0 0 0 0 ...
..$ SB1 : int [1:15] 0 1 1 1 0 1 1 1 1 1 ...
..$ SB2 : int [1:15] 0 0 0 0 1 1 1 1 0 1 ...
..$ SB3 : int [1:15] 0 0 0 0 0 0 0 0 0 0 ...
..$ SB4 : int [1:15] 0 0 0 0 0 0 0 0 0 0 ...
..$ SB5 : int [1:15] 0 0 0 1 1 0 0 0 0 1 ...
..$ SB6 : int [1:15] 0 1 0 1 1 1 1 0 1 0 ...
..$ SB7 : int [1:15] 0 0 0 0 1 0 0 0 0 0 ...

Source

See fraction.subtraction.data for more information about the data source.

See also

References

Chen, Y., Liu, J., Xu, G. and Ying, Z. (2015). Statistical analysis of Q-matrix based diagnostic classification models. Journal of the American Statistical Association, 110(510), 850-866.

de la Torre, J. (2009). DINA model parameter estimation: A didactic. Journal of Educational and Behavioral Statistics, 34, 115-130.

de la Torre, J. (2011). The generalized DINA model framework. Psychometrika, 76, 179-199.

de la Torre, J., & Douglas, J. A. (2004). Higher-order latent trait models for cognitive diagnosis. Psychometrika, 69, 333-353.

de la Torre, J., & Douglas, J. A. (2008). Model evaluation and multiple strategies in cognitive diagnosis: An analysis of fraction subtraction data. Psychometrika, 73, 595-624.

Henson, R. A., Templin, J. T., & Willse, J. T. (2009). Defining a family of cognitive diagnosis models using log-linear models with latent variables. Psychometrika, 74, 191-210.

Huo, Y., & de la Torre, J. (2014). Estimating a cognitive diagnostic model for multiple strategies via the EM algorithm. Applied Psychological Measurement, 38, 464-485.