Test-specific and Item-specific Entropy for Latent Class Models

Computes test-specific and item-specific entropy as test-diagnostic criteria of cognitive diagnostic models (Asparouhov & Muthen, 2014).

entropy.lca(object)

# S3 method for entropy.lca
summary(object, digits=2,  ...)

Arguments

object
Object of class din, gdina or mcdina. For the summary method, it is the result of entropy.lca.

digits
Number of digits to round

...
Further arguments to be passed

Value

A list with the data frame entropy as an entry.

References

Asparouhov, T. & Muthen, B. (2014). Variable-specific entropy contribution. Technical Appendix. http://www.statmodel.com/7_3_papers.shtml

See also

Examples

#############################################################################
# EXAMPLE 1: Entropy for DINA model
#############################################################################

data(sim.dina, package="CDM")
data(sim.qmatrix, package="CDM")

# fit DINA Model
mod1 <- CDM::din( sim.dina, q.matrix=sim.qmatrix, rule="DINA")
summary(mod1)
# compute entropy for test and items
emod1 <- CDM::entropy.lca( mod1 )
summary(emod1)

if (FALSE) {
#############################################################################
# EXAMPLE 2: Entropy for polytomous GDINA model
#############################################################################

data(data.pgdina, package="CDM")

dat <- data.pgdina$dat
q.matrix <- data.pgdina$q.matrix

# pGDINA model with "DINA rule"
mod1 <- CDM::gdina( dat, q.matrix=q.matrix, rule="DINA")
summary(mod1)

# compute entropy
emod1 <- CDM::entropy.lca( mod1 )
summary(emod1)

#############################################################################
# EXAMPLE 3: Entropy for MCDINA model
#############################################################################

data(data.cdm02, package="CDM")

dat <- data.cdm02$data
q.matrix <- data.cdm02$q.matrix

# estimate model with polytomous atribute
mod1 <- CDM::mcdina( dat, q.matrix=q.matrix )
summary(mod1)
# computre entropy
emod1 <- CDM::entropy.lca( mod1 )
summary(emod1)
}