Wald Statistic for Item Fit of the DINA and ACDM Rule for GDINA Model

This function tests with a Wald test for the GDINA model whether a DINA or a ACDM condensation rule leads to a sufficient item fit compared to the saturated GDINA rule (de la Torre & Lee, 2013). The Wald test is accompanied by the RMSEA fit and weighted and unweighted distance measures (wgtdist, uwgtdist), see Details (compare Ma, Iaconangelo, & de la Torre, 2016).

gdina.wald(object)

# S3 method for gdina.wald
summary(object, digits=3,
    vars=c("X2", "p", "sig", "RMSEA", "wgtdist"),  ...)

Arguments

object
A fitted gdina model

digits
Number of digits after decimal used for rounding.

vars
Vector including variables which should be displayed in summary. See the output stats.

...
Further arguments to be passed

Details

Let \(P_j( \alpha _l)\) the estimated item response function for the GDINA model and \(\hat{P}_j( \alpha _l)\) the item response model for the approximated model (DINA, DINO or ACDM). The unweighted distance uwgtdist as a measure of misfit is defined as $$uwgtdist=\frac{1}{2^K} \sum_l ( P_j( \alpha _l) - \hat{P}_j( \alpha _l) )^2$$ The weighted distance wgtdist measures the discrepancy with respected to the probabilities \(w_l=P( \alpha_l)\) of estimated skill classes $$wgtdist=\sum_l w_l (P_j( \alpha _l) - \hat{P}_j( \alpha _l) )^2$$

Value

stats

Data frame with Wald statistic for every item, corresponding p values and a RMSEA fit statistic

References

de la Torre, J., & Lee, Y. S. (2013). Evaluating the Wald test for item-level comparison of saturated and reduced models in cognitive diagnosis. Journal of Educational Measurement, 50, 355-373.

Ma, W., Iaconangelo, C., & de la Torre, J. (2016). Model similarity, model selection, and attribute classification. Applied Psychological Measurement, 40(3), 200-217.

See also

Examples

if (FALSE) {
#############################################################################
# EXAMPLE 1: Wald test for DINA simulated data sim.dina
#############################################################################

data(sim.dina, package="CDM")
data(sim.qmatrix, package="CDM")

# Model 1: estimate GDINA model
mod1 <- CDM::gdina( sim.dina, q.matrix=sim.qmatrix,  rule="GDINA")
summary(mod1)

# perform Wald test
res1 <- CDM::gdina.wald( mod1 )
summary(res1)
# -> results show that all but one item fit according to the DINA rule

# select some output
summary(res1, vars=c("wgtdist", "p") )
}