Print Method for Objects of Class summary.din

S3 method to print objects of the class summary.din.

# S3 method for summary.din
print(x, ...)

Arguments

x
A required object of class summary.din, obtained from a call to the function summary.din (through generic function summary).

...
Optional parameters to be passed to or from other methods will be ignored.

Details

The print method prints the summary information about objects of the class din computed by summary.din, which are the item discriminations indices, the most frequent skill classes and the model information criteria AIC and BIC. Specific summary information details such as individual items with their discrimination index can be accessed through assignment (see ‘Examples’).

Value

If the argument x is of required type, print.summary.din prints the summary information in ‘Details’, and invisibly returns x.

See also

Examples

##
## (1) examples based on dataset fractions.subtraction.data
##

## In particular, accessing detailed summary through assignment
mod <- CDM::din(data=CDM::fraction.subtraction.data,
          q.matrix=CDM::fraction.subtraction.qmatrix, rule="DINA")
smod <- summary(mod)
str(smod)